+++ router/pay_via_payment_details.js
@@ -15,6 +15,13 @@ const subscribeToPayViaDetails = require('./subscribe_to_pay_via_details');
     [max_fee]: <Maximum Fee Tokens To Pay Number>
     [outgoing_channel]: <Pay Out of Outgoing Channel Id String>
     [pathfinding_timeout]: <Time to Spend Finding a Route Milliseconds Number>
+    routes: [[{
+      [base_fee_mtokens]: <Base Routing Fee In Millitokens String>
+      [channel]: <Standard Format Channel Id String>
+      [cltv_delta]: <CLTV Blocks Delta Number>
+      [fee_rate]: <Fee Rate In Millitokens Per Million Number>
+      public_key: <Forward Edge Public Key Hex String>
+    }]]
     [timeout_height]: <Maximum Expiration CLTV Timeout Height Number>
     tokens: <Tokens To Pay Number>
   }
@@ -65,6 +72,7 @@ module.exports = (args, cbk) => {
           max_fee: args.max_fee,
           outgoing_channel: args.outgoing_channel,
           pathfinding_timeout: args.pathfinding_timeout,
+          routes: args.routes,
           timeout_height: args.timeout_height,
           tokens: args.tokens,
         });
@@ -76,7 +84,11 @@ module.exports = (args, cbk) => {
             return cbk([503, 'UnexpectedErrorPayingViaPaymentDetails', {err}]);
           }
 
-          if (!!res.failed && !!res.is_pathfinding_timeout) {
+          if (!!res.failed && !!res.failed.is_invalid_payment) {
+            return cbk([503, 'PaymentRejectedByDestination']);
+          }
+
+          if (!!res.failed && !!res.failed.is_pathfinding_timeout) {
             return cbk([503, 'PaymentAttemptsTimedOut']);
           }